+++ src/validate_submission.py
@@ -2,22 +2,17 @@
 
 import argparse
 import sys
-import re
-import logging
 from pathlib import Path
-from glob import glob
-import csv
+import inspect
 
-from directory_schema.errors import DirectoryValidationErrors
-
-from ingest_validation_tools.validator import (
-    validate, validate_metadata_tsv, validate_data_path, TableValidationErrors
-)
+from ingest_validation_tools.error_report import ErrorReport
+from ingest_validation_tools.submission import Submission
 from ingest_validation_tools import argparse_types
+from ingest_validation_tools.argparse_types import ShowUsageException
 from ingest_validation_tools.globus_utils import get_globus_connection_error
 
 
-def parse_args():
+def make_parser():
     parser = argparse.ArgumentParser(
         description='''
 Validate a HuBMAP submission, both the metadata TSVs, and the datasets,
@@ -63,16 +58,28 @@ Typical usecases:
         help='A list of type / metadata.tsv pairs '
         f'of the form "{expected_type_metadata_form}".')
 
-    # Should there be logging?
+    # How should output be formatted?
 
-    log_levels = ['DEBUG', 'INFO', 'WARN']
-    parser.add_argument(
-        '--logging', type=str,
-        metavar='LOG_LEVEL',
-        help=f'Logging level: One of {log_levels}',
-        choices=log_levels,
-        default='WARN')
+    error_report_methods = [
+        name for (name, type) in inspect.getmembers(ErrorReport)
+        if name.startswith('as_')
+    ]
+    parser.add_argument('--output', choices=error_report_methods,
+                        default='as_text')
 
+    parser.add_argument('--add_notes', action='store_true',
+                        help='Append a context note to error reports.')
+
+    return parser
+
+
+# We want the error handling inside the __name__ == '__main__' section
+# to be able to show the usage string if it catches a ShowUsageException.
+# Defining this at the top level makes that possible.
+parser = make_parser()
+
+
+def parse_args():
     args = parser.parse_args()
     if not any([
         args.local_directory,
@@ -80,132 +87,42 @@ Typical usecases:
         args.globus_origin_directory,
         args.type_metadata
     ]):
-        raise ValidationException(
-            parser.format_usage() +
-            'At least one argument is required')
+        raise ShowUsageException('At least one argument is required')
 
     return args
 
 
 def main():
     args = parse_args()
-    logging.basicConfig(level=args.logging)
 
     globus = args.globus_url or args.globus_origin_directory
     if globus:
         error_message = get_globus_connection_error(globus['origin'])
         if error_message:
-            raise ValidationException(error_message)
+            raise ShowUsageException(error_message)
 
-        raise ValidationException('TODO: Globus not yet supported')
+        raise ShowUsageException('TODO: Globus not yet supported')
         # TODO: mirror directory to local cache.
 
+    submission_args = {'add_notes': args.add_notes}
     if args.local_directory:
-        logging.info(f'Validating {args.local_directory}')
-        messages = _validate_submission_directory_messages(
-            args.local_directory)
-
+        submission_args['directory_path'] = Path(args.local_directory)
     if args.type_metadata:
-        messages = []
-        for type_path in args.type_metadata:
-            logging.info(f'Validating {type_path}')
-            messages += _validate_metadata_tsv_messages(
-                type=type_path['type'],
-                metadata_path=type_path['path']
-            )
-    print('\n'.join(messages))
-
-    return 1 if messages else 0
-
-
-def _validate_metadata_tsv_messages(type, metadata_path):
-    try:
-        validate_metadata_tsv(type=type, metadata_path=metadata_path)
-        logging.info('PASS')
-        return []
-    except TableValidationErrors as e:
-        logging.warning('FAIL')
-        return [str(e)]
-
-
-def _validate_data_path_messages(type, data_path):
-    logging.info(f'Validating {type} {data_path}')
-    try:
-        validate_data_path(type=type, data_path=data_path)
-        logging.info('PASS')
-        return []
-    except DirectoryValidationErrors as e:
-        logging.warning('FAIL')
-        return [str(e)]
-
-
-def _validate_submission_directory_messages(submission_directory):
-    metadata_glob = submission_directory + '/*-metadata.tsv'
-    metadata_tsvs = glob(metadata_glob)
-    if not metadata_tsvs:
-        raise ValidationException(f'Nothing matched {metadata_glob}')
-    messages = []
-    for tsv_path in metadata_tsvs:
-        dir_type = re.match(r'(.+)-metadata\.tsv$', Path(tsv_path).name)[1]
-        table_type = dir_type.split('-')[0]
-        messages += _validate_metadata_tsv_messages(table_type, tsv_path)
-
-        with open(tsv_path) as f:
-            rows = list(csv.DictReader(f, dialect='excel-tab'))
-            if not rows:
-                raise ValidationException(f'{tsv_path} is empty')
-            for row in rows:
-                full_data_path = Path(submission_directory) / row['data_path']
-                messages += _validate_data_path_messages(
-                    dir_type, full_data_path)
-
-    return messages
-
-
-class ValidationException(Exception):
-    # Throw this when there it a problem with the validation process
-    # (not just that validation failed) an you don't want a stack trace.
-    pass
-
-
-def _print_message(
-        dir, type,
-        periods=False, skip_data_path=False):
-    # Doctests choke on blank lines: periods=True replaces with "." for now.
-    try:
-        validate(dir, type, skip_data_path=skip_data_path)
-        logging.info('PASS')
-        return 0
-    except DirectoryValidationErrors as e:
-        message = str(e)
-        if periods:
-            message = re.sub(r'\n(\s*\n)+', '\n.\n', message).strip()
-        # End user just wants a name, and doesn't care about refs.
-        message = re.sub(r"\$ref: '#/definitions/(\w+)'", r'\1', message)
-        # Ad hoc rewrites: Perhaps move these up to the library?
-        message = message.replace(
-            'fails this "oneOf" check',
-            'should be one of these')
-        message = message.replace(
-            'fails this "contains" check',
-            'should contain')
-
-        print(message)
-        logging.warning('FAIL')
-        return 1
-    except TableValidationErrors as e:
-        message = str(e)
-        if periods:
-            message = re.sub(r'\n(\s*\n)+', '\n.\n', message).strip()
-        print(message)
-        logging.warning('FAIL')
-        return 2
+        submission_args['override_tsv_paths'] = {
+            pair['type']: pair['path'] for pair in args.type_metadata
+        }
+    submission = Submission(**submission_args)
+    errors = submission.get_errors()
+    report = ErrorReport(submission.get_errors())
+    print(getattr(report, args.output)())
+    return 1 if errors else 0
 
 
 if __name__ == "__main__":
     try:
         exit_status = main()
-    except ValidationException as e:
+    except ShowUsageException as e:
+        print(parser.format_usage(), file=sys.stderr)
         print(e, file=sys.stderr)
-        sys.exit(2)
+        exit_status = 2
     sys.exit(exit_status)