+++ test.sh
@@ -16,14 +16,9 @@ start src-doctests
 find src | grep '\.py$' | xargs python -m doctest
 end src-doctests
 
-# start pytest
-# pytest -vv
-# end pytest
-
-start fixtures
-echo 'NOTE: There will be warnings below, generated from fixtures which should cause warnings!'
-./test-fixtures.py
-end fixtures
+start examples
+./test-examples.sh
+end examples
 
 start generate
 for TYPE in $(ls docs | grep -v README.md); do # Ignore README and just get subdirectories